Updated App.rb
Updated app.rb to be snake_case variables. For readability reasons. Added features where you can input the selector and it finds it and tests the web application for you. Will be updating certain variable names for simplicity sake, as well as adding methods/functions to make it not only run quicker but also make it easier to read.